modularizing localstorage operations
